16-bit indices (8-bit should be enough)

+++ VulkanRenderer/src/Mesh.h
@@ -43,7 +43,7 @@ class Mesh
 {
 public:
     Mesh(const std::vector<glm::vec3>& positions, const std::vector<glm::vec3>& normals,
-        const std::vector<glm::vec2>& uvs, const std::vector<u32>& indices,
+        const std::vector<glm::vec2>& uvs, const std::vector<u16>& indices,
         const BoundingSphere& boundingSphere,
         const std::vector<Meshlet> meshlets);
 
@@ -53,7 +53,7 @@ public:
     const std::vector<glm::vec3>& GetPositions() const { return m_Positions; }
     const std::vector<glm::vec3>& GetNormals() const { return m_Normals; }
     const std::vector<glm::vec2>& GetUVs() const { return m_UVs; }
-    const std::vector<u32>& GetIndices() const { return m_Indices; }
+    const std::vector<u16>& GetIndices() const { return m_Indices; }
 
     void SetVertexBufferOffset(i32 offset) { m_VertexBufferOffset = offset; }
     void SetIndexBufferOffset(u32 offset) { m_IndexBufferOffset = offset; }
@@ -67,7 +67,7 @@ private:
     std::vector<glm::vec3> m_Positions;
     std::vector<glm::vec3> m_Normals;
     std::vector<glm::vec2> m_UVs;
-    std::vector<u32> m_Indices;
+    std::vector<u16> m_Indices;
 
     i32 m_VertexBufferOffset;
     u32 m_IndexBufferOffset;
